Reorganize server modules
- Move logger into a separate module
- Move shutdown handlers and registration into a separate module
- Move startup logic into a separate module
- Update Rollup config to ensure early registration of shutdown handlers
